refactor: rename it to VectorLegend
