+++ sites/geohub/src/components/controls/VectorLegend.svelte
